pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mountain-trainer"
version = "0.1.0"
description = "Iterative LoRA training for Mount Rainier visibility detection"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "astral>=3.2",
    "ipython>=9.0.0",
    "ipywidgets>=8.1.5",
    "nbclassic>=1.2.0",
    "notebook>=7.3.2",
    "matplotlib>=3.10.8",
    "metar>=1.11.0",
    "opencv-python-headless>=4.13.0.92",
    "peft>=0.18.1",
    "playwright>=1.49.1",
    "pytest>=9.0.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "timm>=1.0.24",
    "tomli>=2.4.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "typer>=0.24.0",
]

[project.scripts]
training = "train.scheduler:app"
collect = "collect.collector:app"
classify = "collect.classifier:cli_wrapper"

[tool.setuptools]
packages = ["train", "collect"]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "tomli-w>=1.2.0",
]
